requirements.txt
shodan
requests
python3-nmap
